+++ src/platform/packages/shared/response-ops/recurring-schedule-form/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-browser",
+  "id": "@kbn/response-ops-recurring-schedule-form",
+  "owner": "@elastic/response-ops",
+  "group": "platform",
+  "visibility": "shared"
+}
