+++ src/platform/packages/shared/response-ops/recurring-schedule-form/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "extends": "../../../../../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "target/types",
+    "types": [
+      "jest",
+      "node",
+      "@emotion/react/types/css-prop",
+      "@kbn/ambient-ui-types"
+    ]
+  },
+  "include": [
+    "**/*.ts",
+    "**/*.tsx"
+  ],
+  "exclude": [
+    "target/**/*"
+  ],
+  "kbn_references": [
+    "@kbn/rrule",
+    "@kbn/es-ui-shared-plugin",
+    "@kbn/i18n",
+  ]
+}
