+++ src/platform/packages/shared/response-ops/recurring-schedule-form/utils/month_day_date.test.ts
@@ -1,0 +1,17 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import moment from 'moment';
+import { monthDayDate } from './month_day_date';
+
+describe('monthDayDate', () => {
+  test('should parse the month and date', () => {
+    expect(monthDayDate(moment('2023-03-23'))).toEqual('March 23');
+  });
+});
