+++ x-pack/platform/plugins/shared/alerting/public/pages/maintenance_windows/helpers/convert_from_maintenance_window_to_form.ts
@@ -8,11 +8,14 @@
 import moment from 'moment';
 import { Frequency } from '@kbn/rrule';
 import { has } from 'lodash';
-import type { FormProps, RecurringScheduleFormProps } from '../components/schema';
+import { getInitialByWeekday } from '@kbn/response-ops-recurring-schedule-form/utils/get_initial_by_weekday';
+import type {
+  RecurrenceFrequency,
+  RecurringSchedule,
+} from '@kbn/response-ops-recurring-schedule-form/types';
+import { RecurrenceEnd } from '@kbn/response-ops-recurring-schedule-form/constants';
+import type { FormProps } from '../components/schema';
 import type { RRuleParams, MaintenanceWindow } from '../../../../common';
-import type { MaintenanceWindowFrequency } from '../constants';
-import { EndsOptions } from '../constants';
-import { getInitialByWeekday } from './get_initial_by_weekday';
 
 export const convertFromMaintenanceWindowToForm = (
   maintenanceWindow: MaintenanceWindow
@@ -38,14 +41,14 @@ export const convertFromMaintenanceWindowToForm = (
 
   const rRule = maintenanceWindow.rRule;
   const isCustomFrequency = isCustom(rRule);
-  const frequency = rRule.freq as MaintenanceWindowFrequency;
+  const frequency = rRule.freq as RecurrenceFrequency;
   const ends = rRule.until
-    ? EndsOptions.ON_DATE
+    ? RecurrenceEnd.ON_DATE
     : rRule.count
-    ? EndsOptions.AFTER_X
-    : EndsOptions.NEVER;
+    ? RecurrenceEnd.AFTER_X
+    : RecurrenceEnd.NEVER;
 
-  const recurringSchedule: RecurringScheduleFormProps = {
+  const recurringSchedule: RecurringSchedule = {
     frequency: isCustomFrequency ? 'CUSTOM' : frequency,
     interval: rRule.interval,
     ends,
