+++ x-pack/platform/plugins/shared/alerting/tsconfig.json
@@ -82,7 +82,8 @@
     "@kbn/core-deprecations-server",
     "@kbn/response-ops-rule-params",
     "@kbn/datemath",
-    "@kbn/elastic-assistant-common"
+    "@kbn/elastic-assistant-common",
+    "@kbn/response-ops-recurring-schedule-form"
   ],
   "exclude": [
     "target/**/*"
